requirements.txt
pandas==2.0.1
openai==0.27.0
python-dotenv==0.21.0
numpy==1.24.2
faiss-cpu==1.7.2
langchain==0.0.113 